Extract layout definitions to well defined objects
