+++ packages/block-editor/src/components/block-list/layout.js
@@ -3,25 +3,14 @@
  */
 import { createContext, useContext } from '@wordpress/element';
 
+/**
+ * Internal dependencies
+ */
+import { getLayoutType } from '../../layouts';
+
 export const defaultLayout = { type: 'default' };
 
 const Layout = createContext( defaultLayout );
-
-function appendSelectors( selectors, append ) {
-	// Ideally we shouldn't need the `.editor-styles-wrapper` increased specificity here
-	// The problem though is that we have a `.editor-styles-wrapper p { margin: reset; }` style
-	// it's used to reset the default margin added by wp-admin to paragraphs
-	// so we need this to be higher speficity otherwise, it won't be applied to paragraphs inside containers
-	// When the post editor is fully iframed, this extra classname could be removed.
-
-	return selectors
-		.split( ',' )
-		.map(
-			( subselector ) =>
-				`.editor-styles-wrapper ${ subselector } ${ append }`
-		)
-		.join( ',' );
-}
 
 /**
  * Allows to define the layout.
@@ -35,62 +24,12 @@ export function useLayout() {
 	return useContext( Layout );
 }
 
-function DefaultLayoutStyle( { selector, layout = {} } ) {
-	const { contentSize, wideSize } = layout;
-
-	let style =
-		!! contentSize || !! wideSize
-			? `
-				${ appendSelectors( selector, '> *' ) } {
-					max-width: ${ contentSize ?? wideSize };
-					margin-left: auto !important;
-					margin-right: auto !important;
-				}
-
-				${ appendSelectors( selector, '> [data-align="wide"]' ) }  {
-					max-width: ${ wideSize ?? contentSize };
-				}
-
-				${ appendSelectors( selector, '> [data-align="full"]' ) } {
-					max-width: none;
-				}
-			`
-			: '';
-
-	style += `
-		${ appendSelectors( selector, '> [data-align="left"]' ) } {
-			float: left;
-			margin-right: 2em;
-		}
-
-		${ appendSelectors( selector, '> [data-align="right"]' ) } {
-			float: right;
-			margin-left: 2em;
-		}
-	`;
-
-	return <style>{ style }</style>;
-}
-
-function FlexLayoutStyle( { selector } ) {
-	return (
-		<style>{ `${ selector } {
-		display: flex;
-		column-gap: 0.5em;
-		align-items: center;
-	}` }</style>
-	);
-}
-
 export function LayoutStyle( { layout = {}, ...props } ) {
 	const type = layout.type ?? 'default';
+	const layoutType = getLayoutType( type );
 
-	if ( type === 'default' ) {
-		return <DefaultLayoutStyle layout={ layout } { ...props } />;
-	}
-
-	if ( type === 'flex' ) {
-		return <FlexLayoutStyle layout={ layout } { ...props } />;
+	if ( layoutType ) {
+		return <layoutType.save layout={ layout } { ...props } />;
 	}
 
 	return null;
